updateref: Recognize locking errors when committing transactions
When a reference is locked already, then git-update-ref(1) will fail and
tell us. This information is getting parsed by Gitaly and gets returned
to the caller via a special `AlreadyLockedError` so that they can handle
this case more gracefully.

The regular expression we use to parse the error message is lacking
though: while the error may happen in the `prepare` phase, it can also
happen in the `commit` phase when the caller skips explicit preparation
of the transaction. And as Git will prefix the phase to the error, we
need to recognize both of them. Right now, we only recognize `prepare`
errors though.

Fix the issue by extending the regular expression to also detect
`commit` phase errors.