ref: Change error code for locked refs in DeleteRefs to Aborted
When trying to delete a reference that is already locked we return a
structured error in DeleteRefs with `codes.FailedPrecondition`. This
fails the litmus test documented in `grpc/codes` though:

> A litmus test that may help a service implementor in deciding
> between FailedPrecondition, Aborted, and Unavailable:
>
>  (a) Use Unavailable if the client can retry just the failing call.
>  (b) Use Aborted if the client should retry at a higher-level
>      (e.g., restarting a read-modify-write sequence).
>  (c) Use FailedPrecondition if the client should not retry until
>      the system state has been explicitly fixed. E.g., if an "rmdir"
>      fails because the directory is non-empty, FailedPrecondition
>      should be returned since the client should not retry unless
>      they have first fixed up the directory by deleting files from it.
>  (d) Use FailedPrecondition if the client performs conditional
>      REST Get/Update/Delete on a resource and the resource on the
>      server does not match the condition. E.g., conflicting
>      read-modify-write on the same resource.

FailedPrecondition should be returned only when the client should not
retry, but given that we are looking at a locking issue that is usually
caused by concurrent calls holding the lock this is wrong. Instead,
Aborted is a better fit as a retry should typically work alright.

Refactor the code to do so.