git: Add logic to detect whether git-cat-file(1) supports `-Z`
We have recently backported a set of patches that add a new `-Z` flag to
git-cat-file(1), which is part of Git v2.41.0.gl1. If set, it causes the
command to print lines that are NUL instead of newline terminated, which
is required to make its output unambiguous when the queries themselves
contain newlines.

As the change isn't yet part of any release we can only conditionally
enable this new mode. Prepare for this by adding the logic to detect
whether Git supports this flag or not.