git: Update to Git v2.41.0.gl1
We have recently upstreamed a new `-Z` option for git-cat-objects(1)
that will cause it to NUL-terminate its output. This is required so that
it becomes possible to request revisions from it that contain embedded
newlines while using one of the `--batch` modes.

We have backported this patch series into Git v2.41.0.gl1, so let's
upgrade to that version so that we can start using this option.