catfile: Refuse input with newlines for old Git versions
While the catfile package can now correctly handle queries with embedded
newlines, this support requires Git v2.41.0.gl1 or v2.42.0 and newer.
For older versions of Git that don't yet support the `-Z` flag for the
git-cat-file(1) command such requests will still fail.

The failure mode with embedded newlines can potentially be quite bad.
Because the git-cat-file(1) process may be cached, and because valid
responses can be forged, it is feasible to poison the cache by putting a
process back into it that still has pending output. The consequence is
that the next request that reuses such a cached process may see the
wrong result.

In general this isn't as bad as it sounds as the cache is per-user. But
it's still a hard-to-debug issue that we should try hard to avoid. So
let's refuse any requests that contain a newline altogether in the case
where the Git version that is in use is too old.

Changelog: fixed