catfile: Fix parsing object info for missing objects containing newlines
With the preceding commit, we have fixed the general case for requests
where the requested revision contains newlines by starting to make use
of the `-z` flag of git-cat-file(1). But while succesful requests work
as expected now, unsuccessful requests for a missing object will still
fail. This is because the output of git-cat-file(1) repeats the input
query:

```
$ printf 'foo\nbar' | git cat-file --batch -z
foo
bar missing
```

This output format is unparsable in general and thus causes the request
to fail.

To fix this we have upstreamed a new `-Z` flag for git-cat-file(1) that
will not only NUL-terminate the input, but also the output. So let's fix
the bug by using this new flag in case we detect that the current Git
version supports it.

Changelog: fixed