catfile: Prepare tests for upcoming changes
We're about to amend our catfile tests so that they support both newline
and NUL terminated input. This will require us to use some options for
the `read` builtin which are Bash constructs. But given that we use a
POSIX shell in our tests this will not work.

Furthermore, we will need to detect the Git version in our tests and
thus require a Gitaly configuration in order to set up the Git command
factory.

Prepare for these changes by using a portable Bash shebang and by
injecting the `config.Cfg` when constructing the intercepted object
queue.