repository service: Stop validating vanity repo in BackupRepository
The vanity repository is used to generate paths to store backups since
we want to store backups using the relative path that gitlab-rails
knows, not the rewritten replica path. As such the vanity repository is
only expected to exist in praefect so there is no way for gitaly to
properly validate this other than simply checking that the fields are
set.

Changelog: fixed