git: Implement new function to validate reference names
While we already provide `git.CheckRefNameFormat()` to verify the format
of reference names, this function relies on git-check-ref-format(1) to
do the verification for us. It is thus not usable as a general purpose
function as it is simply too expensive to run once per reference we are
to verify.

As a first step towards a fix this commit introduces a new function that
performs thorough validation of reference names inside Gitaly itself.
The function aims to be a faithful reimplementation of what Git does and
catch the exact same edge cases.

Subsequent commits will convert callsites of `CheckRefNameFormat()` to
instead use this new function.