Run TestListLastCommitsForTree subtests sequentially
TestListLastCommitsForTree sets up a single repository that is shared
between all of the subtests. The subtests however write more objects into
the repository. This means the state of the repository is unclear when
the tests starts and may change even during the test.

This is prone to cause flakes with transaction management as these writes
happen concurrently without any synchronization. Fix this by running the
subtests sequentially.

Running the tests in parallel with a single repository being set up makes
pretty much no difference to the run time, so do this simpler fix instead
of setting up completely separate test state for each subtest.