conflict: Add argument to append newlines in `Resolve()`
The `Resolve` function is used to programmatically resolve conflicts
during merges. If there is a newline to the end of the file and there is
conflict in said region, then git ends up absorbing the newline of the
file. This is a bug in git and was confirmed on the mailing list:
https://lore.kernel.org/git/CAOLa=ZRumrpfG8FxQQG=Q8tGvxEapMvOx6SZyRPh0GSpn5iuUg@mail.gmail.com/

We need to work around since the git2go flow keeps the newline intact.
So let's add a new argument to the `Resolve` function, wherein clients
would indicate if they want a newline appended to the end of the
resolution, the checking of the source file is left to the client since
this functions primary use-case is to only resolve conflicts and
shouldn't worry about rails affiliated logic.