updateref: Drop use of Ruby-style error
When we have migrated the Operations service into Gitaly, many of the
RPCs were embedding errors into a successful response via a special
field. These errors were in fact consumed and often also displayed by
Rails, which meant that we were not in a position to actually change
them. This led to a bunch of error messages that don't really carry a
lot of information.

We have since migrated most of the RPCs to instead use structured errors
and are thus free to change those errors to be more meaningful. One such
instance is the dreaded "Could not update $ref. Please refresh and try
again."  error message. This error can indeed be caused by a multitude
of different errors, but what the actual root cause is is quite unclear.

Refactor the code and get rid of this error, instead propagating the
actual root cause to the caller. This greatly clarifies the reason why a
reference update has failed and should help us debug such cases better
in the future.

Note that there is once instance left with `UserRevert()` where we have
to return the old-style error as it is still getting embedded into a
successful response. This commit simply inlines that error and adds a
TODO comment on top of it.