repository service: Hande restore skipped errors
This is another edge-case caused by not differentiating between
non-existent and empty repositories. In this case we need to handle not
being able to find a backup bundle in which case we may skip, depending
on AlwaysCreate. Ultimately this would be fixed by
https://gitlab.com/gitlab-org/gitlab/-/issues/357044

Until then, it is useful to respond in a way that isn't successful and
isn't a full blown failure.