repository: Detect locking issues in WriteRef
When WriteRef fails because it cannot lock the reference that is to be
updated it returns an internal error. Given that this is quite a likely
error scenario though and that we already have the infrastructure in
place to detect such issues we can do better though and return a more
telling error.

Refactor the RPC to instead return an explicit error with metadata. This
should make it easier for us to detect lock contention better going
forward.