structerr: Provide convenience function to extract metadata from errors
Callers that wish to extract error metadata from a generic error need to
manually extract the `structerr.Error` first and then call `Metadata()`
on it, which is a bit cumbersome. Furthermore, we're about to extend the
logic to also allow for any other error types to carry metadata via an
interface, which would require them to also be on the lookout for any
such interface types.

Refactor the code and pull out the logic to extract error metadata into
a standalone `structerr.ExtractMetadata()` function. Given a generic
error, it will return all error metadata that is contained in the error
chain. This makes it easy to extend the logic going forward.

Adjust existing callsites to use this new function.