structerr: Allow addding metadata to any error type
In order to add metadata to an error one must create a `structerr.Error`
right now. This is a bit roundabout though in the case where you already
have a custom error type that you wish to add metadata to. While you can
play games and implement an `Unwrap()` function on your error type that
returns a `structerr.Error` with metadata, this is indeed quite tedious
and roundabout.

Instead, provide a new `structerr.ErrorMetadater` interface that can be
implemented by custom error types. This allows them to attach metadata
to the custom error type directly via a `ErrorMetadata()` function.