gitaly-backup: Add -server-side option to create sub-command
This option triggers gitaly-backup to use the new BackupRepository RPC
to create backups. These backups are streamed from the gitaly node
directly to the object-storage configured in config.backup.go_cloud_url.

Tests added for the command are smoke tests as the backup logic itself
is tested on backup.Manager.

Changelog: added