Replace limithandler's stats by generic customfields
Limithandler has a dedicated utility to store gRPC log fields. This
utility suite consists of a recorder, a FieldsProducer, and two
interceptors to plug in gRPC server. customfields can provide the same
service. It was built for general purposes. Hence, there is no reason to
keep both of them. This commit removes limithandler's dedicated stats
handler in favor of customfields.