Introduce more metrics for spawn token
This commit adds three more metrics to monitor spawn token:
- `gitaly_spawn_forking_time_seconds`: Histogram of time waiting for
  spawn tokens.
- `gitaly_spawn_waiting_time_seconds`: Histogram of actual forking time
  after spawn tokens are acquired.
- `gitaly_spawn_token_queue_length`: The current length of the queue
  waiting for spawn tokens.

This commit also removes
`gitaly_command_spawn_token_acquiring_seconds_total` metric. That metric
is a counter which sums all the time waiting for spawn token. It breaks
down by grpc_service then cmd. That breakdown doesn't make sense,
because spawn token acquisition is process-global. The waiting time
doesn't correlate to the command or grpc_service. In addition, counter
is not a good type for this measurement. It is replaced by the new
metrics.

Changelog: added