Support starting a transaction with a non-existent repository
Beginning a transaction currently fails if the repository does not
exist. This is generally fine for RPCs that would attempt to operate
on the repository by doing reference updates for example. The handler
would anyway error out if the repsitory doesn't exist. Some RPCs
however should run even if the repository doesn't exist. For example,
RepositoryExists returns whether the repository exists or not in its
response. The various repository creating RPCs also should run even
if the repository doesn't exist. To accommodate for those RPCs, this
commit allows starting a transaction even if the repository doesn't
exist. The RPC handlers anyway check for repository existence, so it's
fined to call them without a repository.

Tests are updated to verify the repository doesn't exist throgh a
separate assertion where previously they relied on a not found error
being returned.