Ignore relative paths when asserting error metadata
Some errors include the repository's relative path in the error
metadata and assert the metadata for equality. This is not ideal
as the relative path is an internal storage detail and should not
be asserted against elsewhere as we are now doing in handler tests.
These tests will all fail when the relative path is rewritten. Address
this by overriding the relative path field when constructing intecepted
metadata. This way the key's existence in the error metadata is asserted
but not the actual contents of the key.

It would be better to fix the problem in a less magical way but there's
a large number of tests that are asserting the relative paths for equality.
Fix these tests for now by overriding the value. We can later fix the tests
one by one to not assert the relative path to remove this.