Migrate objects to the repository's configured object directory
Our quarantine helpers always migrate the objects from the quarantine
to '<repo.git>/objects'. This works if the repository hasn't been
configured with a different object directory which may be the case
if the repository is already quarantined. If that's the case, we should
instead migrate the objects to the configured object directory. Support
this by considering the configured object directory before defaulting
to migrating to 'objects'.

While double quarantines should be avoided, they allow for migrating
all existing RPCs to a new quarantining approach without having to
touch all of the RPCs in the initial migration.