diff: Refactor tests for RawDiff
The tests for RawDiff are kind of weird: instead of verifying the
expected result, they only verify that reapplying the computed diff
would again result in the original tree. This is quite roundabout and
does not help to actually verify what matters, namely that the diff that
is getting generated matches both ours and the callers expectations.

Refactor the tests to generate test data at runtime so that it becomes
easy for us to derive the exact expected state. This also helps to make
the tests ready for the SHA256 object format.