gittest: Deduplicate default commit time
In order to arrive at deterministic dates for our test-generated commits
we have a hardcoded default commit time that we use in `WriteCommit()`.
This default commit date is hardcoded in multiple places though,
duplicating the information.

Unify all of the sites into the `DefaultCommitTime` variable.

Note that this requires us to adapt one of our testscases for
`gitaly-git2go`'s cherry-pick subcommand. Previously we were both
passing in and expecting an empty timezone name, but with the
refactoring this changes so that we do in fact pass in a non-empty name.
This is seemingly not reflected in Git2go, causing the test. The issue
is fixed by overriding the expected timezone name again.