gitaly-git2go: Drop "apply" subcommand
The "apply" subcommand has never really been used in production due to a
bug in libgit2 that causes certain patches to misapply. As we have never
been able to get the fix upstreamed, we were barred from using libgit2
here and instead still continue to use a worktree and git-apply(1). As
we have essentially deprecated libgit2, this is also not going to change
either.

We still had a dependency on the "apply" subcommand in our tests for
RawDiff and RawPatch. These have been refactored now though, removing
the last usage of the command. So let's drop the implementation of the
"apply" subcommand altogether.