git2go: Rename "commit.go" to "errors.go"
In the preceding commit we have removed the implementation of the
"commit" subcommand. But as "commit.go" still contains a bunch of shared
error types we cannot remove that file yet, but it's name does not match
its contents anymore.

Rename it to "errors.go" to better reflect what it contains.