git2go: Move commit actions to UserCommitFiles
The commit actions are still declared in the `internal/git2go` package
even though they have nothing to do with Git2go anymore. Let's move them
to their only site of usage, which is next to the `UserCommitFiles()`
implementation.

Note that we also rename the structure from `Action` to `commitAction`.
This is done to avoid local name conflicts.