Add only the configured object directory as an alternate
Quarantining works by pointing the object directory to the quarantine
directory and setting the original object directroy as an alternate.
This way all object writes will end up in the quarantine but the
original objects remain available through the alternate.

Apply is currently unconditionally setting 'objects' as an alternate.
If a repository that has already been quarantined is quarantined again,
this leads to adding duplicate 'objects' entries in the alternates. Fix
this by adding 'objects' as an alternate only if the repository didn't
have another object directory already configured.