conflicts: Add new field `skip_content_parsing`
The `ListConflictFiles` RPC merges the provided tree-ish objects and
returns the conflicted files along with their contents. Sometimes, we
don't really care about the content of the conflicted files (for e.g.
merge compatibility checks in CI/CD pipelines).

Parsing file contents and streaming them is a laborious process and
slows down the RPC. We should allow the clients to decide if they really
need the contents of the conflicted files. The new fields will help us
implement the functionality where the RPC would simply check for
conflicts and return the conflicted filenames without the contents.

While we're updating the proto with the new field. We also add missing
documentation for `ListConflictFilesRequest`.