conflicts: Implement `SkipContentParsing` in `ListConflictFiles`
In the previous commit, we introduced the new field `SkipContentParsing`
in the proto definition of `ListConflictFilesRequest`. The field will be
used to skip conflicted file's content parsing and streaming.

Let's implement the usage of the field by simply skipping the content
parsing and streaming section within the `ListConflictFiles` RPC. Also
add a test to test for this new functionality.