repository service: Allow RestoreRepository to restore latest backup
Restoring currently only works if every repository that is being
restored has a backup. This is a problem because backups cannot
differentiate between empty repositories and non-existent repositories.
See https://gitlab.com/gitlab-org/gitlab/-/issues/357044

In the meantime we can mitigate the effect of this by restoring backups
how backup.rake currently restores, that is by only ever restoring from
the latest backup. This works because it then has to query what the
latest backup is and so can tell that there are actually no backups,
rather than erroring on a non-existent backup.

Changelog: changed