repository: Support SHA256 repositories for `GetSnapshot()`
When `GetSnapshot()` is invoked for a SHA256 repository, the RPC
succeeds, but the resulting snapshot does not contain any objects. This
is because the regular expressions used to add objects to the snapshot
do not recognize SHA256 hashes.

Update the regular expressions to support SHA256. Also remove the build
tag from `snapshot_test.go` preventing inclusion with SHA256 tests.