limiter: Handle adaptive concurrency limit's watcher timeout
When the adaptive concurrency calculator polls for the backoff events,
it's possibile that the watcher is timeout. It's a signal for a backoff
event. However, a single timeout is too vauge. The calculator should
treat a timeout as a backoff event only if there were N consecutive
timoeuts already. This commit introduces a timeout counter and handle
timeout events accordingly.