operations: Implement UserRevert in pure git
When reverting a commit, we actually hope to undo all the changes
brought in by the commit. The changes brought in by the commit are diffs
from the commit's parent to the commit itself, so the diffs from the
commit itself to its parent are exactly the opposite changes.

By merging in a commit's parent and using the commit itself as merge
base, what we get would be the commit reverted. And for commits having
no parents at all, we can assume their parent is an empty tree.

Inspired-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>
Inspired-by: Jiuyang Xie <jiuyang.xjy@alibaba-inc.com>
Inspired-by: Toon Claes <toon@gitlab.com>
Helped-by: Han Xin <hanxin.hx@bytedance.com>
Signed-off-by: Xing Xin <xingxin.xx@bytedance.com>
Changelog: changed