ci: Require manual execution of Rails spec jobs
With 9b1ef44c, the GitLab Rails suite of tests was introduced to the
Gitaly pipeline and required to pass in order for merges to be accepted.
Unfortunately, the Rails spec jobs are frequently flaky thus blocking
merges.

Set the `start-rails-specs` CI job to only run if executed manually.
Also remove the rules from `rails-specs` and `rails-specs-cleanup` so
they execute automatically when `start-rails-specs` is triggered.