backup: Add RemoveAllRepositories to Strategy interface
Now that all implementations have this method, it can be added to the
Strategy interface. This will allow us to switch between the actual
backup implementation in backup.Manager and the server-side adapter in
backup.ServerSideAdapter.

Now that this method is part of the interface, the request type is moved
to be close to the interface instead of backup.Manager