protoregistry: Move package into `internal/grpc` for more general use
The `protoregistry` package is currently only used by Praefect, which is
why it lives in `internal/praefect` right now. We're about to extend its
usage to also cover Gitaly though.

Move the package into `internal/grpc` so that it becomes clearer that it
is generic and can be reused across components.