metadatahandler: Expose method scope and operation
We log a bunch of gRPC information via the `metadatahandler` package. In
several contexts though we have been wishing for rough classifiers of
gRPC methods so that we can for example easily filter for all methods
that mutate a repository.

Introduce two new fields:

    - "grpc.meta.method_operation" denotes whether the RPC is an
      accessor-, mutator- or maintenance-style RPC.

    - "grpc.meta.method_scope" denotes whether the RPC is repository- or
      storage-scoped.

With this it becomes a lot simpler to filter full categories of RPCs.

Changelog: added