commit: Make CommitStats SHA256-compatible
The CommitStats RPC calculates commit statistics for changes introduced
with a commit. One edge case is when we inspect the root commit which
doesn't have any parents. In order to treat this case the same as the
other cases we use the empty tree object ID as the old commit such that
all blobs are being treated as new. We still use the hardcoded SHA1
object ID though, which makes this edge case incompatible with SHA256
repositories.

Detect the object hash at runtime which makes it compatible with SHA256
repositories. Enable tests to run with SHA256.