Allow including unreachable objects in transaction
Some RPCs in Gitaly, such as UserSquash, write objects into the repository
without making them reachable. The TransactionManager only includes objects
in the logged pack file that are reachable. This creates a problem for those
RPCs as the objects they attempt to write do not get committed into the
repository. This commit enables committing unreachable objects in a transaction.
The problematic RPCs can include specific objects in the pack even if they
are not reachable from the references so they get committed as part of the
transaction.