backup: Extract file creation test helper
Testing restores requires that a lot of different files be created in
specific locations and this has resulted in tests that have a lot of
boilerplate that tries to generate paths, make directory trees and write
files. Often creating a lot of confusingly named variables in the
process.

To help these tests be more understandable a helper has been extracted
testhelper.WriteFiles. This helper simply writes out a collection of
files to disk in a more declarative way, such that the filesystem
structure is more visible.