Support updating a reference multiple times within a transaction
Transaction manager currently only supports setting the reference
updates of a transaction once. This makes it difficult to gather the
reference updates from the reference transaction hook as there may
be multiple invocations of the hook within a single transaction. With
the current way things are, only the updates from the last invocation
of the hook would be committed.

This commit improves upon this by supporting multiple updates to the
references withing a single transaction. This enables then recording
the updates from the hooks which may be invoked multiple times. As
each transaction will only commit the reference updates at the end,
the updates withing a transaction are compacted from the start state
to the end state. So updates like 'oid-1 -> oid-2 -> oid-3' are
ultimately comitted as 'oid-1 -> oid-3' as the intermediate states
within a transaction don't make a difference when it comes to the
final result.