Record initial values of updated references
The reference transaction hook has an issue where for both force updates
and reference deletions the old OID is zero OID. It's thus not possible
to distinguish them without resolving the references value in the 'prepare'
stage of the reference transaction hook. If we resolve the value in the
'prepare' phase, we also don't want to stage an update for a reference
before the 'commit' phase as the reference may not actually get updated.
This commit introduces RecordInitialReferenceValues method on the transaction
that can be used to record a reference's initial value as seen in the
'prepare' phase prior to any updates without actually staging the
reference for an update. This recorded initial values can then be used as
the old OID when ultimately staging the references for an update through
UpdateReferences.