ci: Fix rails spec job run outside of an MR
Currently the `rails-specs` jobs rely on the `CI_MERGE_REQUEST_IID` and
`CI_MERGE_REQUEST_SOURCE_BRANCH_SHA` variables to execute. However, if a
pipeline is run outside of the context of an MR then these variables are
undefined and the jobs will fail.

To resolve this, rely on the pipeline iid to uniquely identify the
branch created in 'gitlab-org/gitlab'.